gemini_chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai
python-dotenv
python-multipart
pydantic
pydantic-core
typing-extensions
openai
packaging